Java: combine test cases; add test for StaplerResponse.forward
